Add folders, break up some large files
