+++ AuroraEditor/Base/Editor/CodeEditorView/UI/CodeView.swift
@@ -33,17 +33,14 @@ struct MessageInfo {
 ///
 typealias MessageViews = [LineInfo.MessageBundle.ID: MessageInfo]
 
-// MARK: -
-// MARK: AppKit version
-
 /// `NSTextView` with a gutter
 ///
 class CodeView: NSTextView { // swiftlint:disable:this type_body_length
 
     // Delegates
-    fileprivate let codeViewDelegate = CodeViewDelegate()
-    fileprivate let codeLayoutManagerDelegate = CodeLayoutManagerDelegate()
-    fileprivate var codeStorageDelegate: CodeStorageDelegate
+    let codeViewDelegate = CodeViewDelegate()
+    let codeLayoutManagerDelegate = CodeLayoutManagerDelegate()
+    var codeStorageDelegate: CodeStorageDelegate
 
     // Subviews
     var gutterView: GutterView?
@@ -483,288 +480,6 @@ class CodeView: NSTextView { // swiftlint:disable:this type_body_length
 
 }
 
-class CodeViewDelegate: NSObject, NSTextViewDelegate {
-
-    // Hooks for events
-    //
-    var textDidChange: ((NSTextView) -> Void)?
-    var selectionDidChange: ((NSTextView) -> Void)?
-
-    // MARK: NSTextViewDelegate protocol
-
-    func textDidChange(_ notification: Notification) {
-        guard let textView = notification.object as? NSTextView else { return }
-
-        textDidChange?(textView)
-    }
-
-    func textViewDidChangeSelection(_ notification: Notification) {
-        guard let textView = notification.object as? NSTextView else { return }
-
-        selectionDidChange?(textView)
-    }
-}
-
-// MARK: -
-// MARK: Shared code
-
-extension CodeView {
-
-    /// Update the layout of the specified message view if its geometry got invalidated by
-    /// `CodeTextContainer.lineFragmentRect(forProposedRect:at:writingDirection:remaining:)`.
-    ///
-    fileprivate func layoutMessageView(
-        identifiedBy id: UUID
-    ) {
-        guard let codeLayoutManager = layoutManager as? CodeLayoutManager,
-              let codeStorage = textStorage as? CodeStorage,
-              let codeContainer = optTextContainer,
-              let messageBundle = messageViews[id]
-        else { return }
-
-        if messageBundle.geometry == nil {
-            let glyphRange = codeLayoutManager.glyphRange(
-                forBoundingRect: messageBundle.lineFragementRect,
-                in: codeContainer
-            ),
-                charRange = codeLayoutManager.characterRange(
-                    forGlyphRange: glyphRange,
-                    actualGlyphRange: nil
-                ),
-                lineRange = (codeStorage.string as NSString).lineRange(for: charRange),
-                lineGlyphs = codeLayoutManager.glyphRange(
-                    forCharacterRange: lineRange,
-                    actualCharacterRange: nil
-                ),
-                usedRect = codeLayoutManager.lineFragmentUsedRect(
-                    forGlyphAt: glyphRange.location,
-                    effectiveRange: nil
-                ),
-                lineRect = codeLayoutManager.boundingRect(forGlyphRange: lineGlyphs, in: codeContainer)
-
-            // Compute the message view geometry from the text layout information
-            let geometry = MessageView.Geometry(
-                lineWidth: messageBundle.lineFragementRect.width - usedRect.width,
-                lineHeight: messageBundle.lineFragementRect.height,
-                popupWidth: (codeContainer.size.width - MessageView.popupRightSideOffset) * 0.75,
-                popupOffset: lineRect.height + 2
-            )
-            messageViews[id]?.geometry = geometry
-
-            // Configure the view with the new geometry
-            messageBundle.view.geometry = geometry
-            if messageBundle.view.superview == nil {
-
-                // Add the messages view
-                addSubview(messageBundle.view)
-                let topOffset = textContainerOrigin.y + messageBundle.lineFragementRect.minY,
-                    topAnchorConstraint = messageBundle.view.topAnchor.constraint(equalTo: self.topAnchor,
-                                                                                  constant: topOffset)
-                let leftOffset = textContainerOrigin.x + messageBundle.lineFragementRect.maxX,
-                    rightAnchorConstraint = messageBundle.view.rightAnchor.constraint(equalTo: self.leftAnchor,
-                                                                                      constant: leftOffset)
-                messageViews[id]?.topAnchorConstraint = topAnchorConstraint
-                messageViews[id]?.rightAnchorConstraint = rightAnchorConstraint
-                NSLayoutConstraint.activate([topAnchorConstraint, rightAnchorConstraint])
-
-            } else {
-
-                // Update the messages view constraints
-                messageViews[id]?.topAnchorConstraint?.constant = messageBundle.lineFragementRect.minY
-                messageViews[id]?.rightAnchorConstraint?.constant = messageBundle.lineFragementRect.maxX
-
-            }
-        }
-    }
-
-    /// Adds a new message to the set of messages for this code view.
-    ///
-    func report(message: Located<Message>) {
-        guard let messageBundle = codeStorageDelegate.add(message: message) else { return }
-
-        updateMessageView(for: messageBundle, at: message.location.line)
-    }
-
-    /// Removes a given message. If it doesn't exist, do nothing. This function is quite expensive.
-    ///
-    func retract(message: Message) {
-        guard let (messageBundle, line) = codeStorageDelegate.remove(message: message) else { return }
-
-        updateMessageView(for: messageBundle, at: line)
-    }
-
-    /// Given a new or updated message bundle, update the corresponding message view appropriately.
-    /// This includes covering the two special cases, where we create a new view or we remove a view for\
-    ///  good (as its last message was deleted).
-    private func updateMessageView(for messageBundle: LineInfo.MessageBundle, at line: Int) {
-        guard let charRange = codeStorageDelegate.lineMap.lookup(line: line)?.range else { return }
-
-        removeMessageViews(withIDs: [messageBundle.id])
-
-        // If we removed the last message of this view, we don't need to create a new version
-        if messageBundle.messages.isEmpty { return }
-
-        // TODO: CodeEditor needs to be parameterised by message theme
-        let theme = Message.defaultTheme
-
-        let messageView = StatefulMessageView.HostingView(messages: messageBundle.messages,
-                                                          theme: theme,
-                                                          geometry: MessageView.Geometry(lineWidth: 100,
-                                                                                         lineHeight: 15,
-                                                                                         popupWidth: 300,
-                                                                                         popupOffset: 16),
-                                                          fontSize: font?.pointSize ?? OSFont.systemFontSize),
-            principalCategory = messagesByCategory(messageBundle.messages)[0].key,
-            colour = theme(principalCategory).colour
-
-        messageViews[messageBundle.id] = MessageInfo(view: messageView,
-                                                     lineFragementRect: CGRect.zero,
-                                                     geometry: nil,
-                                                     colour: colour)
-
-        // We invalidate the layout of the line where the message belongs as their may be
-        // less space for the text now and because the layout process for the text fills
-        // the `lineFragmentRect` property of the above `MessageInfo`.
-        optLayoutManager?.invalidateLayout(forCharacterRange: charRange, actualCharacterRange: nil)
-        self.optLayoutManager?.invalidateDisplay(forCharacterRange: charRange)
-        gutterView?.invalidateGutter(forCharRange: charRange)
-    }
-
-    /// Remove the messages associated with a specified range of lines.
-    ///
-    /// - Parameter onLines: The line range where messages are to be removed. If `nil`,\
-    /// all messages on this code view are to be removed.
-    func retractMessages(onLines lines: Range<Int>? = nil) {
-        var messageIds: [LineInfo.MessageBundle.ID] = []
-
-        // Remove all message bundles in the line map and collect their ids for subsequent view removal.
-        for line in lines ?? 1..<codeStorageDelegate.lineMap.lines.count {
-
-            if let messageBundle = codeStorageDelegate.messages(at: line) {
-
-                messageIds.append(messageBundle.id)
-                codeStorageDelegate.removeMessages(at: line)
-
-            }
-
-        }
-
-        // Make sure to remove all views that are still around if necessary.
-        if lines == nil { removeMessageViews() } else { removeMessageViews(withIDs: messageIds) }
-    }
-
-    /// Remove the message views with the given ids.
-    ///
-    /// - Parameter ids: The IDs of the message bundles that ought to be removed. If `nil`, remove all.
-    ///
-    /// IDs that do not have no associated message view cause no harm.
-    ///
-    fileprivate func removeMessageViews(withIDs ids: [LineInfo.MessageBundle.ID]? = nil) {
-
-        for id in ids ?? [LineInfo.MessageBundle.ID](messageViews.keys) {
-
-            if let info = messageViews[id] { info.view.removeFromSuperview() }
-            messageViews.removeValue(forKey: id)
-
-        }
-    }
-
-    /// Ensure that all message views are in their collapsed state.
-    ///
-    func collapseMessageViews() {
-        for messageView in messageViews {
-            messageView.value.view.unfolded = false
-        }
-    }
-
-}
-
-class CodeContainer: NSTextContainer {
-
-    override func lineFragmentRect(forProposedRect proposedRect: CGRect,
-                                   at characterIndex: Int,
-                                   writingDirection baseWritingDirection: NSWritingDirection,
-                                   remaining remainingRect: UnsafeMutablePointer<CGRect>?)
-    -> CGRect {
-        let calculatedRect = super.lineFragmentRect(forProposedRect: proposedRect,
-                                                    at: characterIndex,
-                                                    writingDirection: baseWritingDirection,
-                                                    remaining: remainingRect)
-
-        guard let codeView = textView as? CodeView,
-              let codeStorage = layoutManager?.textStorage as? CodeStorage,
-              let delegate = codeStorage.delegate as? CodeStorageDelegate,
-              let line = delegate.lineMap.lineOf(index: characterIndex),
-              let oneLine = delegate.lineMap.lookup(line: line),
-              characterIndex == oneLine.range.location    // we are only interested in the first line fragment of a line
-        else { return calculatedRect }
-
-        // On lines that contain messages, we reduce the width of the available line fragement rect such that there is
-        // always space for a minimal truncated message (provided the text container is wide enough to accomodate that).
-        if let messageBundleId = delegate.messages(at: line)?.id,
-           calculatedRect.width > 2 * MessageView.minimumInlineWidth {
-
-            codeView.messageViews[messageBundleId]?.lineFragementRect = calculatedRect
-            codeView.messageViews[messageBundleId]?.geometry = nil // invalidate the geometry
-
-            // To fully determine the layout of the message view,
-            // typesetting needs to complete for this line; hence, we defer
-            // configuring the view.
-            DispatchQueue.main.async { codeView.layoutMessageView(identifiedBy: messageBundleId) }
-
-            return CGRect(origin: calculatedRect.origin,
-                          size: CGSize(width: calculatedRect.width - MessageView.minimumInlineWidth,
-                                       height: calculatedRect.height))
-
-        } else { return calculatedRect }
-    }
-}
-
-/// Customised layout manager for code layout.
-///
-class CodeLayoutManager: NSLayoutManager {
-
-    weak var gutterView: GutterView?
-
-    override func processEditing(for textStorage: NSTextStorage,
-                                 edited editMask: TextStorageEditActions,
-                                 range newCharRange: NSRange,
-                                 changeInLength delta: Int,
-                                 invalidatedRange invalidatedCharRange: NSRange) {
-        super.processEditing(for: textStorage,
-                             edited: editMask,
-                             range: newCharRange,
-                             changeInLength: delta,
-                             invalidatedRange: invalidatedCharRange)
-
-        // NB: Gutter drawing must be asynchronous, as the glyph generation that may be triggered in that process,
-        //     is not permitted until the enclosing editing block has completed; otherwise, we run into an internal
-        //     error in the layout manager.
-        if let gutterView = gutterView {
-            Dispatch.DispatchQueue.main.async { gutterView.invalidateGutter(forCharRange: invalidatedCharRange) }
-        }
-
-        // Remove all messages in the edited range.
-        if let codeStorageDelegate = textStorage.delegate as? CodeStorageDelegate,
-           let codeView = gutterView?.textView as? CodeView {
-
-            codeView.removeMessageViews(withIDs: codeStorageDelegate.lastEvictedMessageIDs)
-
-        }
-    }
-}
-
-class CodeLayoutManagerDelegate: NSObject, NSLayoutManagerDelegate {
-
-    func layoutManager(_ layoutManager: NSLayoutManager,
-                       didCompleteLayoutFor textContainer: NSTextContainer?,
-                       atEnd layoutFinishedFlag: Bool) {
-        guard let layoutManager = layoutManager as? CodeLayoutManager else { return }
-
-        if layoutFinishedFlag { layoutManager.gutterView?.layoutFinished() }
-    }
-}
-
 /// Common code view actions triggered on a selection change.
 ///
 func selectionDidChange<TV: TextView>(_ textView: TV) {
@@ -782,32 +497,6 @@ func selectionDidChange<TV: TextView>(_ textView: TV) {
        location > 0,
        let matchingBracketRange = codeStorage.matchingBracket(forLocationAt: location - 1, in: charRange) {
         textView.showFindIndicator(for: matchingBracketRange)
-    }
-}
-
-extension NSLayoutManager {
-
-    /// Enumerate the fragment rectangles covering the characters located on the line with the given character index.
-    ///
-    /// - Parameters:
-    ///   - charIndex: The character index determining the line whose rectangles we want to enumerate.
-    ///   - block: Block that gets invoked once for every fragement rectangles on that line.
-    ///
-    func enumerateFragmentRects(forLineContaining charIndex: Int, using block: @escaping (CGRect) -> Void) {
-        guard let text = textStorage?.string as NSString? else { return }
-
-        let currentLineCharRange = text.lineRange(for: NSRange(location: charIndex, length: 0))
-
-        if currentLineCharRange.length > 0 {  // all, but the last line (if it is an empty line)
-
-            let currentLineGlyphRange = glyphRange(forCharacterRange: currentLineCharRange, actualCharacterRange: nil)
-            enumerateLineFragments(forGlyphRange: currentLineGlyphRange) { (rect, _, _, _, _) in block(rect) }
-
-        } else {                              // the last line if it is an empty line
-
-            block(extraLineFragmentRect)
-
-        }
     }
 }
